+++ vllm/model_executor/layers/attention_backends/paged_attn.py
@@ -118,25 +118,8 @@ class PagedAttention:
         key_cache: torch.Tensor,
         value_cache: torch.Tensor,
         input_metadata: InputMetadata,
-        num_heads: int,
-        num_kv_heads: int,
         alibi_slopes: Optional[torch.Tensor],
     ) -> torch.Tensor:
-        num_queries_per_kv = num_heads // num_kv_heads
-        if num_kv_heads != num_heads:
-            # TODO(woosuk): Use MQA/GQA kernels for higher performance.
-            query = query.view(query.shape[0], num_kv_heads,
-                                num_queries_per_kv,
-                                query.shape[-1])
-            key = key[:, :,
-                        None, :].expand(key.shape[0], num_kv_heads,
-                                        num_queries_per_kv,
-                                        key.shape[-1])
-            value = value[:, :,
-                            None, :].expand(value.shape[0],
-                                            self.num_kv_heads,
-                                            self.num_queries_per_kv,
-                                            value.shape[-1])
         output = torch.empty_like(query)
         context_attention_fwd(
             query,
